add basic test
